Final version without pathcalculator

+++ src/pathfinder/Node.cpp
@@ -1,7 +1,7 @@
 #include "Node.h"
 
 
-Node::Node(int id, int x_position, int y_position)
+Node::Node(int id, float x_position, float y_position)
 {
 	_id = id;
 	_x_position = x_position;
@@ -18,6 +18,6 @@ bool Node::operator==(const Node &n) const {
 bool Node::operator!=(const Node &n) const {
     return _id != n._id;
 }
-int Node::getXPosition() { return _x_position; }
-int Node::getYPosition() { return _y_position; }
+float Node::getXPosition() { return _x_position; }
+float Node::getYPosition() { return _y_position; }
